Split JSONAPI modules into separate source files
